Use pure modules for dynamic imports
