+++ packages/dev/loaders/src/glTF/2.0/Extensions/KHR_animation_pointer.ts
@@ -5,7 +5,7 @@
 export * from "./KHR_animation_pointer.types";
 export * from "./KHR_animation_pointer.pure";
 
-import "./KHR_animation_pointer.data";
-
+import { _RegisterKHRAnimationPointerData } from "./KHR_animation_pointer.data.pure";
 import { RegisterKHR_animation_pointer } from "./KHR_animation_pointer.pure";
+_RegisterKHRAnimationPointerData();
 RegisterKHR_animation_pointer();
