+++ packages/dev/loaders/src/glTF/2.0/Extensions/pure.ts
@@ -45,3 +45,4 @@ export * from "./KHR_node_selectability.pure";
 export * from "./KHR_node_hoverability.pure";
 export * from "./ExtrasAsMetadata.pure";
 export * from "./KHR_interactivity/pure";
+export * from "./KHR_animation_pointer.data.pure";
