+++ packages/public/@babylonjs/loaders/package.json
