+++ scripts/treeshaking/side-effects-manifest/loaders/glTF.json
@@ -21,7 +21,7 @@
         "glTF/2.0/Extensions/ExtrasAsMetadata.ts": ["top-level-call"],
         "glTF/2.0/Extensions/ExtrasAsMetadata.types.ts": ["declare-module"],
         "glTF/2.0/Extensions/KHR_animation_pointer.data.ts": ["top-level-call"],
-        "glTF/2.0/Extensions/KHR_animation_pointer.ts": ["bare-import", "top-level-call"],
+        "glTF/2.0/Extensions/KHR_animation_pointer.ts": ["top-level-call"],
         "glTF/2.0/Extensions/KHR_animation_pointer.types.ts": ["declare-module"],
         "glTF/2.0/Extensions/KHR_draco_mesh_compression.ts": ["top-level-call"],
         "glTF/2.0/Extensions/KHR_draco_mesh_compression.types.ts": ["declare-module"],
